fix: move budget alert logic to team.py file, and other minor cleanups
(cherry picked from commit 26899f203e3262fd87f5af0dd70884f668102520)
