+++ press/templates/emails/budget_alert.html
@@ -1,31 +1,33 @@
-{% extends "templates/emails/base.html" %} {% block content %}
+{% extends "templates/emails/base.html" %}
+{% import "templates/emails/macros.html" as utils %}
+{% block content %}
 <td class="text-base leading-6 text-gray-900">
 	<div class="p-8 bg-white">
 		<div class="from-markdown">
 			<p>Hello,</p>
 
 			<p>
-				Your Frappe Cloud team "<strong>{{ team_user }}</strong>" cost has exceeded
-				the monthly budget alert limit.
+				Your Frappe Cloud cost for team <strong>{{ team_user }}</strong> has exceeded
+				the monthly budget alert threshold.
 			</p>
 
-			<h3>
+			<h3 style="margin: 5px 0; font-weight: bold;">
 				Budget Alert Details:
 			</h3>
 			<table class="table">
 				<tr>
 					<td>
-						Monthly Budget Limit:
+						Monthly Alert Threshold:
 					</td>
-					<td style="padding: 8px 12px; border: 1px solid #ddd">
-						{{ budget_limit }}
+					<td>
+						{{ alert_threshold }}
 					</td>
 				</tr>
 				<tr>
 					<td>
 						Month-To-Date Invoice Amount:
 					</td>
-					<td style="padding: 8px 12px; border: 1px solid #ddd">
+					<td>
 						{{ invoice_amount }}
 					</td>
 				</tr>
@@ -48,8 +50,7 @@
 			</table>
 
 			<p style="margin: 20px 0">
-				To manage your budget alerts or update your monthly limit, please visit
-				your billing settings in the Frappe Cloud dashboard.
+				To manage your budget alerts, please visit your billing settings in the Frappe Cloud dashboard.
 			</p>
 
 			{{ utils.separator() }}
